+++ contracts/PKPNFT.sol
@@ -178,7 +178,7 @@ contract PKPNFT is
         require(msg.value == mintCost, "You must pay exactly mint cost");
         uint tokenId = _getNextTokenIdToMint(keyType);
         _mintWithoutValueCheck(tokenId, address(this));
-        pkpPermissions.addPermittedAction(tokenId, ipfsCID, new uint[](0));
+        pkpPermissions.addPermittedAction(tokenId, ipfsCID, new uint8[](0));
         _burn(tokenId);
         return tokenId;
     }
@@ -227,7 +227,7 @@ contract PKPNFT is
         onlyOwner
     {
         _mintWithoutValueCheck(tokenId, address(this));
-        pkpPermissions.addPermittedAction(tokenId, ipfsCID, new uint[](0));
+        pkpPermissions.addPermittedAction(tokenId, ipfsCID, new uint8[](0));
         _burn(tokenId);
     }
 
@@ -258,7 +258,7 @@ contract PKPNFT is
         freeMintSigTest(freeMintId, msgHash, v, r, s);
         _mintWithoutValueCheck(tokenId, address(this));
         redeemedFreeMintIds[freeMintId] = true;
-        pkpPermissions.addPermittedAction(tokenId, ipfsCID, new uint[](0));
+        pkpPermissions.addPermittedAction(tokenId, ipfsCID, new uint8[](0));
         _burn(tokenId);
     }
